requirements.txt
numpy>=1.24.0
torch>=2.0.0
pygame>=2.5.0
matplotlib>=3.7.0
pillow>=10.0.0
